update the Dockerfile
We can now safely copy all of the conf dir, since I dropped the toml
file.

This also changes the second stage Docker base image from distroless to
alpine. This takes our image size up from ~25MB to ~30MB, but we get
things like a shell for that little increase, and that will be useful.