switch to `tsgo` and a Go-based build program (#75)
* switch to `tsgo` and a Go-based build program

tsgo is very much pre-release, but it seems to work perfectly for
our TS files, so let's do it. The benefit is that the repo no longer
requires devs to have node to get/upgrade `tsc`. No they'll simply
be able to use the TS compiler, tsgo, through `go tool tsgo`

* remove the old build script