get new Incident number as part of INSERT statement (#100)
...same for Field Report number

previously we were doing a two-step thing, in which we first
queried to find the max number for a given event, then we second
did an insert using that max number + 1. Now I'm doing it all in
one insert statement. The only caveat is that I had to go outside
sqlc to make this work, but it's good to see how simple it can be
to do that.