show login button when root page loads
currently no login/logout button is shown until the client determines
whether there's an authenticated user or not. On a slow connection,
or with a puny server (e.g. staging), it takes quite a while for that
determination to come back. In 90% of cases, a user will be arriving
at this page because they want to log in, so we should just show
the login button until we possibly decide it's not needed.