make JWTs a lot shorter, using bitsets
previously we stored position and team names inside users' JWTs. A JWT
is just a base64-encoded marshalling of the JSON of that data, which
meant that a Ranger with many positions and many teams could easily have
a JWT of over 1 KB. We get into problem territory quickly after that,
because browsers won't sent headers if they're too large.

This change makes the JWT much shorter, by encoding the positions and
teams as their IDs, in a bitset format. This does them in base 16, but
we could go crazier and use base 36 if we want marginally more space
saving in the future (or if the number of positions and teams increases
dramatically).

https://github.com/burningmantech/ranger-ims-go/issues/119