get GetEventAccesses to show permissions for each rule
Mostly I wanted to make it easier for IMS admins to see what each rule
ends up doing.

This involved a bunch of refactoring that leaves UserStore in better
shape than it was in. Now the caching of users happens at a higher
level, with positions and teams merged into that cache.