add an admin server debugging page (#217)
* add an admin server debugging page

this allows us to see build info and lots of runtime information about
the server. There's also a button/API to trigger a garbage collection,
as that helps a lot for viewing runtime metrics (e.g. you'd want to
trigger a GC then see the memory use right after that).

* upgrade harden-runner to appease dependabot

https://github.com/burningmantech/ranger-ims-go/security/dependabot/6