bring back datetime title display in table pages
The stretched-link was causing the td elements to be underneath the
link, and thus the title wasn't displaying.

This tweaks things a bit, so that all of the text in the table is above
the stretched-link anchor. This makes titles show up on hover, and it
allows a user to select the text (e.g. to copy it to clipboard).

I had to make this not happen on mobile though, because a mobile user
would then be confused when they tap on a row but don't get redirected.

https://github.com/burningmantech/ranger-ims-go/issues/220