fix display of hidden incident types (#248)
* fix display of hidden incident types

I recently switched around the APIs to talk about incident types by
their IDs, rather than by their names. That had a side effect of making
hidden incident types vanish completely in the UI from incidents they
had been previously attached to. This change does enough for those types
to still appear in the UI on those incidents, but it stops the hidden
incident types from showing up to be added to new incidents (or to be
filtered by on the incidents page)

* sort incidents table by start time rather than IMS #

now that start time is mutable, IMS # ordering no longer matches start
time ordering. It'll be more intuitive to users that start time be used
as the default sort order.