improve error message when IMS can't connect to S3 (#335)
this has us pass along the server's error message to the web client
when a file upload or download fails. I want it to be clear to users on
playa when this sort of error happens, so that they know they should try
again when internet connectivity is restored.

As part of this, I renamed fetchJsonNoThrow to fetchNoThrow, because
it's really no longer just about handling JSON responses.