don't reset pagination on SSE update
an Operator observed this year than when she was on a search results
page other than 1, she got sent back to page 1 automatically on any
update to any incident (due to the SSE-induced update to the table).
This tweak makes it so that DataTables keeps the user on the page they
were on in that case.