enhance docker-compose use (#458)
the main thing here is that we stop using the cosmtrek/air docker image,
and rather make an image and install air directly. That's preferable
because it solves the problem of cosmtrek/air never being up to date
with the latest Go version (e.g. currently 1.25.2 is out, but comtrek
hasn't released an image with that version yet, so we get blocked on
upgrading to 1.25.2 as well).

This change also does some other cleanups on Docker code.