represent location addresses as strings throughout IMS
The main reason to change this is so that we can harmonize addresses
with the Burning Man API. When a Ranger starts to type an art/camp into
the "location name" field, we'll be able to automatically fill in the
address. The API expresses much more variety of addresses than we could
do with the simple hour:minute@concentric model.

https://github.com/burningmantech/ranger-ims-go/issues/430