feat: add "event groups" for easier permissioning
the one key missing part is that there's no way from the web client to
associate an event with a group. I need to restructure the EditEvents
endpoint in order to make that possible. This PR was getting big enough
already though, so this is something of a progress check-in.

https://github.com/burningmantech/ranger-ims-go/issues/492

https://github.com/burningmantech/ranger-ims-go/issues/260